added project management services to create project, task, subtask an…

+++ Models/InputModels/ProjectTimesheetModel.cs
@@ -11,7 +11,7 @@ namespace TimesheetBE.Models.InputModels
         public Guid? SubTaskId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int PercentageOfCompletion { get; set; }
+        public double PercentageOfCompletion { get; set; }
         public bool Billable { get; set; }
     }
 }
